pyproject.toml
[project]
name = "llomax"
version = "0.1.0"
description = "LLM agent-based pipeline for creating artistic collages from Internet Archive material"
requires-python = ">=3.11"
dependencies = [
    "Pillow",
    "anthropic",
    "httpx",
    "internetarchive>=5.7.2",
    "numpy",
    "onnx",
    "opencv-python",
    "openvino",
    "python-dotenv",
    "segment-anything",
    "torch>=2.0",
    "torchvision",
]

[project.scripts]
llomax = "llomax.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llomax"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff>=0.15.1",
]

[tool.ruff]
line-length = 99

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
